Update python 2.7 patches to build on Mac ARM64.
This backports the final commits to the cpython repository, rather than
the work-in-progress patches that were used previously.

Tested by running the cpython test suite with a native x86-64 build,
a native ARM64 build, and on ARM64 with a cross-compiled build.

Not updating the patch version because we don't yet have this package
built for Mac ARM64 at all.

Bug: 1124119
Change-Id: Ia5279f812e0b599c13e6ee92bce644ad819cde86
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2850059
Commit-Queue: Brian Ryner <bryner@google.com>
Reviewed-by: Owen Rodley <orodley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#40392}