Fix a bug with ARM64 cross builds.
Including the host cpu in _PYTHON_HOST_PLATFORM breaks some logic in
setup.py which expects the platform to be == "Darwin". As a result, it
does not pass the right defines via -D arguments to the compiler,
causing the _ctypes module to use the incorrect code paths for Apple's
system libffi and crash at runtime.

Bug: 1184767
Change-Id: I091039e81dbcefb859529183632a5b762c96fecc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2850061
Reviewed-by: Owen Rodley <orodley@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/master@{#40393}