pinpoint: Initial support for user configurations
This change adds support for reading a YAML config for the Pinpoint CLI.
By default we're looking for an environment variable
`PINPOINT_USER_CONFIG` which should point to a file. To begin with,
we'll add support for setting the defaults of flags. Because flags are
configured later in the subcommands, we are able to specify defaults
lazily.

If we don't find the `PINPOINT_USER_CONFIG` environment, we'll use
$XDG_CONFIG_HOME/pinpoint/config.yaml or equivalent filenames. On Linux
for example, this might be $HOME/.config/pinpoint/config.yaml if
`XDG_CONFIG_HOME` isn't defined.

We use https://golang.org/pkg/os/#UserConfigDir for crafting the
configuration filename.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1202636
Change-Id: I1d58800c26e85f3b007695fa78d7ffc4e8ec5eb3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2850737
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#40424}