pinpoint: Add support for experiment presets
This change introduces the concept of presets, which define a bundle of
flags/options hosted in a file that's a sort of database. It also
integrates with user-specific configuration to allow defining a default
set of presets, which can be user-specific or project specific.

The default preset filename is `.pinpoint-presets.yaml` which would be
searched for in the directory from which the CLI is used. The intent is
to allow projects to define the presets per directory to start with. A
presets file can also be provided as a flag `-presets-file=`, which is a
convenience mechanism to abstract details of which benchmark,
configuration, and story selection mechanisms, and extra args are
defined for a preset. We use the PINPOINT_PRESETS environment variable
first to determine the location of the presets file, which is overridden
by setting the flag.

User-specific configuration also supports defining the default location
for the presets file, which enables defining preset files which
overrides that the default would have been for the flag.

We also add a `config` command which shows what the new default flags
are for when running the CLI tool reading from the user-specific
configuration file.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1202636
Change-Id: I7d3f1f14a9b665a36abcb9209d53a1ed3e8134dc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2855151
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#40519}