pinpoint: Improve conversion and listing support
This change ensures that we're properly handling the case when
comparison mode is empty, which when they're started from the Pinpoint
UI is not `try`.

We're also using the new option 'INPUTS' which will be supported by the
legacy service to provide minimal information (not the full job state)
when giving the list of jobs.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1203942
Change-Id: I3486425db6fc0120a3c338c573eb649f77475fdd
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2858499
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Dean Berris <dberris@chromium.org>
Cr-Commit-Position: refs/heads/master@{#40520}