pinpoint: Add a user agent to requests
This forwards the user agent provided by the client to the legacy
service. The raw arguments are exported to BigQuery as part of a JSON
string (arguments) which allows us to perform some queries to
differentiate by the provided user agent.

We can make this a first-class attribute of the Job model later, but
in the meantime this is sufficient to provide useful information.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1204372
Change-Id: I15e69a29e8b64d9f7a29eabf86752c366191039a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2861646
Commit-Queue: Dean Berris <dberris@chromium.org>
Commit-Queue: Andrew Bennetts <abennetts@google.com>
Auto-Submit: Dean Berris <dberris@chromium.org>
Reviewed-by: Andrew Bennetts <abennetts@google.com>
Cr-Commit-Position: refs/heads/master@{#40555}