pinpoint: Set a user agent for the CLI
This allows us to identify requests made with the Pinpoint CLI. The user
agent by default contains the user's OS and architecture, and a static
date in the identifier. Since we don't do semantic versioning (yet) this
would be sufficient for now just for the purpose of analytics in the
exported BigQuery data on Pinpoint jobs.

R=abennetts@google.com, fancl@chromium.org

Bug: chromium:1203942
Change-Id: Ia98c017ed5b37e3f29ffee6cf214d2ac2e36c194
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2861647
Reviewed-by: Andrew Bennetts <abennetts@google.com>
Commit-Queue: Dean Berris <dberris@chromium.org>
Cr-Commit-Position: refs/heads/master@{#40558}