[cr-audit-commits] Skip robot-owned changes
Some robot-created changes do not have the robot's email as the commit
author. For example, Copybara supports forging the commit author's email
to be the same as the email of the original commit. In this case we need
to look at the owner of the Gerrit change to determine whether the
change is robot-owned or not.

Also remove an unused argument to the `getChangeWithLabelDetails`
function.

Bug: fuchsia:70304
Change-Id: Id2eee6de3e4bc127dc2d93a29d1321cbb718a9b8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2862292
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Xinyu Yang <xinyuoffline@google.com>
Cr-Commit-Position: refs/heads/master@{#40569}