Use LDFLAGS_NODIST for --version-script in CPython build.
The value of LDFLAGS at build time is stored in the resulting Python
distribution and read by distutils to determine LDFLAGS for building
extensions. But this flag (a) is only relevant to the CPython build
itself (b) refers to a path that only exists during the build.

This hasn't mattered up until this point as the resulting CPython
package is only used for running already built wheels. But with the move
to use our CIPD CPython in dockerbuild, we now need to make sure its
LDFLAGS are clean and suitable for building wheels.

Bug: 1176014
Change-Id: Ia4593d7746fcec1935c1d07e849306da6d19beaa
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2870987
Commit-Queue: Owen Rodley <orodley@chromium.org>
Auto-Submit: Owen Rodley <orodley@chromium.org>
Reviewed-by: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/master@{#40626}