Explicitly disable ensurepip for our Python build.
We currently use pip_bootstrap to install pip into our Python packages.
On some platforms ensurepip is already being run, with a different
version of pip. This results in us installing one version of pip on top
of another version, which results in a broken pip installation.

Ideally we'd just use ensurepip and get rid of pip_bootstrap, but that
looks like it'll require some extra work to ensure it works on all
platforms and with cross-compiled builds, so for now lets just disable
it across the board to fix the current brokenness.

Bug: 1205739
Change-Id: Ia7200a088a1368978d88ad77d3cdfa1686fefd40
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2874005
Auto-Submit: Owen Rodley <orodley@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Reviewed-by: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/master@{#40655}