🚝 React Autocomplete: Trigger onChange when Autocomplete loses focus.
For autoSelect: "If `true`, the selected option becomes the value of
the input when the Autocomplete loses focus unless the user chooses
a different option or changes the character string in the input."

Based on the code, the specific behavior for autoSelect with freeSolo
is that if the popup is open and an item is selected, onBlur will
select that option. Otherwise, it will select the value of the input.

https://github.com/mui-org/material-ui/blob/83d7f634ced4c30158695e5791a96c8f3bb85c04/packages/material-ui/src/useAutocomplete/useAutocomplete.js#L829

Change-Id: I66838fd390e80c7e6c31b07a998bd7e904b5497e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2893571
Auto-Submit: Dave Tu <dtu@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Commit-Queue: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#40941}