[dirmd] Update Metadata.imports docs
Now that dirmd is tied to git, "root" means repo root, or in in the
multi-repo case, the root dir of the root repo.

Previously imports were relative to -root flag value which didn't have
to be the repo root. It wasn't great because a DIR_METADATA
author was expected to know how dirmd is called.
Now dirmd respects the repo root, and a DIR_METADATA author no longer
have to think how dirmd is called.

Update the docs that import paths are source-absolute and the root
is the repo root.

Change-Id: Ib8cd20f8a33af148490eba9256c3543f52fc3475
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2894847
Reviewed-by: Josip Sokcevic <sokcevic@google.com>
Commit-Queue: Nodir Turakulov <nodir@chromium.org>
Cr-Commit-Position: refs/heads/master@{#40947}