Fix cross-compiled wheels to build with CIPD-installed cpython.
In order to make sure we use the correct compiler and linker command
lines, install the target cpython as well when cross-compiling, and
point PYTHONPATH at that installation. This will allow distutils to
pick up the target platform _sysconfigdata which is generated at build
time with appropriate compile/link commands for this docker container.

Bug: 1176014
Change-Id: I165abaad010b97bbc7e3586d69afd97629240868
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2891128
Commit-Queue: Brian Ryner <bryner@google.com>
Reviewed-by: Owen Rodley <orodley@chromium.org>
Cr-Commit-Position: refs/heads/master@{#40954}