pinpoint: Add an artifact manifest
This change introduces an artifact manifest which allows us to preserve
some information about the job and which files have been downloaded are
associated with the base and experiment. We'll use this data for when
we're running the analysis over the test result artifacts.

R=abennetts@google.com, fancl@chromium.org, seanmccullough@google.com

Bug: chromium:1204933
Change-Id: I1469fe42fa0f0b5a45626796538fa76f5f058b56
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2894184
Auto-Submit: Dean Berris <dberris@chromium.org>
Commit-Queue: Dean Berris <dberris@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Cr-Commit-Position: refs/heads/master@{#40957}