Don't point CPPFLAGS and LDFLAGS at 3pp deps for the boostrap python.
This unbreaks building cpython at head on linux-amd64.

The issue is that the static libraries in $DEPS_PREFIX/lib are not
built with -fPIC, meaning they cannot be used to build a shared library
module. But when building the bootstrap python, we don't have the full
logic available yet to create the static module list with all of the
appropriate dependencies (that's why we're building the bootstrap
python!).

With this change, the bootstrap python will use the system versions of
openssl and bz2 libraries, which are fine for our purposes, and allow
pip_bootstrap to run successfully.

Bug: 1209116
Change-Id: I793567b583fd3b100548be628560911ad54f166a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2897324
Auto-Submit: Brian Ryner <bryner@google.com>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Cr-Commit-Position: refs/heads/master@{#41002}