tls: Add component updater metadata file
In order for component updater in ash-chrome to forward the
lacros-chrome image, it requires a metadata file. This CL generates that
metadata file, but doesn't provision into a target path.

The users of TLS lacros-chrome provisioning are responsible for copying
to the target paths for ash-chrome to pickup.

BUG=b:189142173
TEST=# ProvisionLacros w/ TLS+TLW services

Change-Id: Iacfc5bf1d0c1a1a50ca603dc5283e2fa301c4098
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2915126
Reviewed-by: Jared Loucks <jaredloucks@google.com>
Reviewed-by: Allen Li <ayatane@chromium.org>
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Cr-Commit-Position: refs/heads/master@{#41232}