Defer SSL default cert initialization until needed.
The old way of doing this causes CoreFoundation to be initialized
at import time on Mac, which spawns threads and is therefore unsafe
if the application later calls os.fork().

Note that there's still a potential race if SSL is initialized prior
to os.fork(), but there's not much we can do about this. This approach
fixes all the broken cases I'm aware of in our codebase.

Bug: 1215121
Change-Id: I9e8279631603ddbd92b1fd110dd728673525c699
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2943752
Commit-Queue: Brian Ryner <bryner@google.com>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Cr-Commit-Position: refs/heads/master@{#41500}