Remove eshwarn@chromium.org from go/src/infra/cmd/shivas/OWNERS
This is an automatically generated changelist based on group memberships and affiliations.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

Change-Id: Ic36abc1bb5d6e6c2e0d0b4cc7eeeb4615f3f7297
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2950466
Reviewed-by: Anushruth  <anushruth@chromium.org>
Commit-Queue: Anushruth  <anushruth@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/master@{#41548}