[lucifer] Add explicit check for socket path length.
On Linux, the length for a socket path is capped at 108 chars
including the null terminator. In order to make the error message
that we get from failing to create a socket more informative,
check the limit up front and fail early with a dedicated error message.

Here is a task that failed due to trying to create a socket path that
was too long.

https://chromeos-swarming.appspot.com/task?id=53c88ccc70c6e110

Here is an excerpt of the message shown:

```
.../abort_sock: bind: invalid argument
```

Here is a quick demonstration that this message really is necessary:

```
package main

import (
        "fmt"
        "net"
        "strings"
)

func main() {
        path := strings.Repeat("a", 200)
        _, err := net.Listen("unixgram", path)
        fmt.Printf("%s\n", err)
}
```

produces the following output

```
listen unixgram aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa: bind: invalid argument
```

BUG=None
TEST=unit tests AND confirm hypothesis about invalid argument message locally

Change-Id: I4e310a518f4c43f3bb8b98a23ab00e31dad46cbe
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2924185
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Reviewed-by: Allen Li <ayatane@chromium.org>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Cr-Commit-Position: refs/heads/master@{#41600}