pinpoint: Add the analyze subcommand
This change defines both the mixin and subcommand for performing
statistical analysis on a job. This also updates the user configuration
documentation to include controlling the defaults analysis and whether
we should return a non-zero exit status when we detect a statistically
significant difference in the experiment.

The new flags for experiments are:

  -analyze-experiment={true|false}
  -check-experiment={true|false}

New configuration options are:

  analyze_experiment={true|false}
  check_experiment={true|false}

For now the output is in YAML, reflecting the structure of the internal
representation used by the command. Following CLs should implement both
human-targeted (text report) and structured (CSV, JSON) output.

R=abennetts@google.com, fancl@chromium.org, seanmccullough@google.com

Bug: chromium:1201147
Change-Id: Ia66d2c628f1faada85fb1be7daeca1ea6b875b24
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2936158
Commit-Queue: Chenlin Fan <fancl@chromium.org>
Reviewed-by: Sean McCullough <seanmccullough@google.com>
Auto-Submit: Dean Berris <dberris@chromium.org>
Cr-Commit-Position: refs/heads/master@{#41658}