Remove eshwarn@google.com from go/src/infra/libs/fleet/OWNERS
This is an automatically generated changelist based on group memberships and affiliations.

To approve this change, CR+1/2 this change and submit. To reject this change, CR -1 this change, and the bot will abandon the change.

Change-Id: I8c0fc37d1a90bc8a71215b3f741fc8ce6ff4dcd0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2957978
Reviewed-by: Anushruth  <anushruth@chromium.org>
Commit-Queue: Anushruth  <anushruth@chromium.org>
Auto-Submit: Owners Cleanup <swarming-tasks@owners-cleanup-prod.google.com.iam.gserviceaccount.com>
Cr-Commit-Position: refs/heads/master@{#41747}