Delete "pre_exec.sh" in our Docker images, and clean up a bit.
This file is executed on startup (if it exists) just before executing
the given command. It 'chown's a bunch of files under /opt/python to the
executing user.

This doesn't seem necessary since these files aren't written to, and
they're already world-readable. And at any rate, we don't use the
built-in Python, we install our own from CIPD.

Not executing it is a big win in startup time for the docker image,
taking it from upwards of 30s down to ~1s.

While updating the image version, get rid of the separate Python 3
template and remove a bunch of redundant code from the template.

NOTRY=true

Bug: 1218302
Change-Id: I87db619a2fb6a630ac532cec704aa96be85f2a4f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2967695
Commit-Queue: Owen Rodley <orodley@chromium.org>
Reviewed-by: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/master@{#41791}