Add a gitiles client for the chromium bootstrapper.
gitiles.Client provides operations for downloading a single file from
gitiles and for resolving a ref to a revision. Downloading single files
will enable the bootstrapper to evaluate the DEPS file and the builder's
properties files, allowing it to set the properties for the bootstrapped
recipe. Resolving a ref to a revision will enable the bootstrapper to
communicate to the bootstrapped recipe the revision of any downloaded
files so that the bootstrapped recipe can sync to the same revisions.

Bug: 1210566
Change-Id: I97aa4e426be2c4f5fd4b5934e09120710e011208
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2947846
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Gary Tong <gatong@chromium.org>
Cr-Commit-Position: refs/heads/master@{#41832}