Avoid a 'cipd resolve' call when fetching source.
Prior to this CL, we first call 'cipd resolve' to check if the source
package exists, and if so, we download it. This is unnecessary, as we
can just try to download it and infer that it doesn't exist if that
fails with a particular error message.

Bug: 1218302
Change-Id: Ibc6f484aa03bf7f6f7c5790dae5668566226a34b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2976165
Commit-Queue: Owen Rodley <orodley@chromium.org>
Auto-Submit: Owen Rodley <orodley@chromium.org>
Reviewed-by: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/master@{#41854}