🚝 Remove user-entered value from autocomplete options.
Having the value listed is always a little awkward,
but it's extra awkward when the field doesn't support
autocomplete and there's no other options.

We have it in case the user needs to custom type something that's a
substring of one of the autocomplete options, but that's a very
uncommon case and there's a workaround by pressing Esc to close the
autocomplete dropdown.

Bug: monorail:5711, monorail:5828
Change-Id: Id3e6fce8a9d7dbb01340c8916301f5dbe5d53afe
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2993539
Auto-Submit: Dave Tu <dtu@chromium.org>
Reviewed-by: Andrew Chang <andrewjc@google.com>
Commit-Queue: Andrew Chang <andrewjc@google.com>
Cr-Commit-Position: refs/heads/master@{#41947}