Report bootstrap failures via the build summary.
A number of errors were updated to surround an object with {} instead of
<> because when those errors are displayed in the summary, all of the
text between the <> disappears due to being interpreted as a tag.
Attempting to HTML escape the summary results in quote characters not
being unescaped in Milo, so changing the characters seems the most
expedient solution.

Bug: 1210566
Change-Id: I20064c8e7f32678fe277c305c7d5436e492b3282
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/2995163
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Cr-Commit-Position: refs/heads/master@{#41972}