Don't run virtualenv.py from inside vpython.
Doing so leads to problems when vpython and infra/bootstrap do not
share the same version of virtualenv. This avoids the requirement
to keep those in lockstep, and is probably more predictable in general.

Tested on Linux and Windows by removing infra/ENV and then running
"gclient runhooks".

Bug: 1219868
Change-Id: Ibd0fc26fd47fab534b1cb1c2d83f0e61fe61e897
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3002762
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Owen Rodley <orodley@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/master@{#42072}