[cloudbuildhelper] Use golang.org/x/tools/go/packages.
We need to traverse Go package tree to bundle *.go source files
for GAE tarballs. We used "go/build" package for this, but it doesn't
work correctly in Go Modules mode.

Use golang.org/x/tools/go/packages instead which is modules-aware and
also significantly faster.

It has a downside: there's no way to tell it to use a different set
of "go1.xx" release tags: it uses `go list` call underneath and thus
always uses the current go release version. This is important when
building GAE tarballs bundles for e.g. "go111" runtime (any runtime
really, since the latest GAE runtime is go115 which is already
stale).

As a partial workaround, use "go/build" just to post-filter *.go
files based on release tags. It is not 100% correct, since if
a discarded file imports a package, this package is still considered
to be a dependency by "go list" (because "go list" call didn't
discard this file).

This situation was modeled in tests, but luckily there are no real
occurrences in all tarballs we have right now. In fact, this CL
results in no changes to existing tarballs whatsoever.

R=iannucci@chromium.org, tandrii@chromium.org
BUG=1057067

Change-Id: I44783dbb4bcb898d289d60c32cb3bd0f71e06f70
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3011349
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Cr-Commit-Position: refs/heads/master@{#42133}