Introduce a reduced interface for the cr-rev gitiles Client.
Re-using the generated gitiles proto client interface results in the
code breaking if new methods are added to the interface. Introducing a
new interface with only the subset of operations that are needed means
that methods can be added to the proto interface without impacting
users.

Bug: 1229211
Change-Id: I0e75097642f870a616f31d0bef2332cc978dc226
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3026650
Commit-Queue: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Cr-Commit-Position: refs/heads/master@{#42335}