Fix 3pp version comparison operators.
The existing operators have a bug where LT or GT will evaluate to true
for the given version, instead of excluding it as expected.

I think this came about due to some confusion with the inverse
operators. For A < B, it's true that !(A >= B), but since we are also
switching the order of the arguments, we need to use B > A.

Change-Id: I08cd56b1fae152e617b54ca429778f47815c9d79
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3035207
Auto-Submit: Brian Ryner <bryner@google.com>
Commit-Queue: Andrii Shyshkalov <tandrii@google.com>
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>
Cr-Commit-Position: refs/heads/master@{#42393}