Fix a bug where we updated the closed_timestamp incorrectly.
In the V3 API, when an IssueDelta doesn't contain a status, it does not
set an old status. The old status is passed into UpdateClosedTimestamp
which passes the old status into MeansOpenInProject that returns "true"
if the status cannot be matched. So when an IssueDelta doesn't update
status, our code thought the issue was open and needed to be closed.

Fixed: monorail:9965
Change-Id: I8f54c362e9069bb5197e0d5c6a4dd320d51c0699
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3042413
Auto-Submit: Andrew Chang <andrewjc@google.com>
Commit-Queue: Dave Tu <dtu@chromium.org>
Reviewed-by: Dave Tu <dtu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#42504}