[go] Don't fail bootstrap if "go install" fails.
It happens often when changing go.mod files in a gcilent
checkout. It is annoying that eval `./env.py` stops working
if something is temporary inconsistent, since "go" in PATH is
needed to fix inconsistencies.

Note that bots no longer depend on thus installed tools and humans,
presumably, will be smart enough to realize something is wrong based
on the error messages in the console.

R=tandrii@chromium.org, iannucci@chromium.org
BUG=1057067

Change-Id: I244c455025e88ee6c6435f6dc6892ae13106fcdb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3046315
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#42551}