Add a field for transmitting reasons for skipping analysis.
When support for gerrit changes is added, changes that affect the
properties should trigger a full rebuild since properties can have
arbitrary effects on the build. This provides a means for the
bootstrapper to communicate to the bootstrapped executable that the
properties were affected so that the bootstrapped executable does not
need to be familiar with any details of the bootstrapping.

Bug: 1210566
Change-Id: I3c5b4c13f3366e47767ad3f8063ce50333e86e7e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3046575
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Stephen Martinis <martiniss@chromium.org>
Cr-Commit-Position: refs/heads/master@{#42559}