Introduce a scope for the errgoup operations.
The new scope enables shadowing the outer value of ctx so that the outer
value can't be used for operations within the errgoup as is currently
happening with some of the log calls. This is an easy mistake to make
and can be hard to spot, so introducing an additional scope will prevent
further slips.

Change-Id: I12628a596d3e564628393e4f7a665c26996fddaf
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3046950
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>
Auto-Submit: Garrett Beaty <gbeaty@chromium.org>
Cr-Commit-Position: refs/heads/master@{#42564}