🚝 Fix issue detail dirty forms handling.
reset() is being called after the properties update but before the DOM
updates. The uncontrolled components use the DOM to get the initial
values when calculating the deltas. So the deltas were using the old
initial values, causing them to erroneously be dirty.

The hacky solution was to ensure that the DOM is up to date in reset()
before calling _processChanges() (which calls _updateIsDirty()).


Bug: monorail:9948
Change-Id: Ied11317baab24967c2a6e7caa48ec02bf4dabbb0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3048839
Commit-Queue: Dave Tu <dtu@chromium.org>
Auto-Submit: Dave Tu <dtu@chromium.org>
Reviewed-by: Tiffany Zhang <zhangtiff@chromium.org>
Cr-Commit-Position: refs/heads/master@{#42584}