[dirmd] Resolve symlinks
In chromeos, generate_location_tags.py is run in the dir with a symlink, instead of the checked out repo. So dirmd complains about failing to process the symlink because it's not under the git root.

This change evaluates symlinks and find the original dirs, and then use the original ones to find the git root.

Bug: b:194513801
Change-Id: Ideb14a213bdb051aa1440d3dca89aae2b3265aa1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3058518
Reviewed-by: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Dirk Pranke <dpranke@google.com>
Commit-Queue: Chan Li <chanli@chromium.org>
Cr-Commit-Position: refs/heads/master@{#42753}