[result_adapter] Do not add the -convey-silent flag.
This flag was added as goconvey was interfering with the the option to
export `go test` results to json, but it was wrong to assume that all
tests would use the goconvey library, and those that don't do not run
the tests as they don't recognize the flag.

As of this patch, the error that originally motivated the use of this
flag is no longer reproducing, so removing the flag should be enough.

Bug: 1234867
Change-Id: I5707eb3e0f969a4cb263994e30e6382c6f48afaa
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3064578
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#42755}