Set OBJCOPY in dockcross containers.
A new version of libffi is needed to fix a bug compiling with
newer gcc versions on linux-mips.

Make sure we choose a dockcross upstream tag that exists for all
of the images. For linux-mipsel, this is simply "latest", as that
port was discontinued prior to the switch to Debian Bullseye.

Bug: 1227833, 1234526
Change-Id: I635a0a9f4c52b354ef01d3b9190a6182b6f9d74f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3058477
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Reviewed-by: Owen Rodley <orodley@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/master@{#42765}