[recovery]: Copy file from remote device to local machine.
This commit is a part of the PARIS project. It provides an
implementation of the method CopyFileFrom of the Access interface.

This method copies the remote source file mentioned in a structure
passed as a function parameter over to the local destination location,
also mentioned in this structure parameter.

BUG=b:193437315
TEST=build, manual testing.

Change-Id: Idad944a29327dbc66aeec7875d7bd5f31837d2d0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3057889
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/master@{#42791}