[recovery]: Copy file from local machine to remote device.
This commit is a part of the PARIS project. It provides an
implementation of the method CopyFileTo of the Access interface.

This method copies the local source file mentioned in a structure
passed as a function parameter over to the remote destination
location, also mentioned in this structure parameter.

BUG=b:193437315
TEST=build, manual testing.

Change-Id: Ic3bc4d5886c67671fcbd872d5b7d13ba078051e6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3069500
Reviewed-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Gregory NISBET <gregorynisbet@google.com>
Commit-Queue: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#43016}