[recovery]: Copy a directory from remote device to local machine.
This commit is a part of the PARIS project. It provides an
implementation of the method CopyDirectoryFrom of the Access
interface.

This method copies the remote source directory mentioned in a
structure passed as a function parameter over to the local destination
location, also mentioned in this structure parameter.

BUG=b:193437315
TEST=build, manual testing.

Change-Id: Ifb40249a4798cf7176a09001dc06518a0d628046
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3088399
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Cr-Commit-Position: refs/heads/main@{#43047}