[recovery]: Copy a directory from local to remote device.
This commit is a part of the PARIS project. It provides an
implementation of the method CopyDirectoryTo of the Access
interface.

This method copies the local source directory mentioned in a structure
passed as a function parameter over to the remote destination
location, also mentioned in this structure parameter.

BUG=b:193437315
TEST=build, manual testing.

Change-Id: Icbf9ce11b8af5f2f0c6020f285864014c6ee4705
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3097312
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#43124}