Per project buildspecs: Add support for multiple projects/watch paths
This CL allows users to invoke the per project buildspec tool as follows:
./foo --projects=galaxy/milkyway,brya/brya --paths=full/buildspecs/,buildspecs/
--min_milestone=94

This change tweaks the specific invocation for the existing use case
(specifically --project becomes --projects and we get rid of --program).

The tool will then scan the appropriate dirs in the external manifest-versions
project and create project-specific buildspecs for each project for each
relevant buildspec. As an early optimization, we only create a project-specific
buildspec if that buildspec does not yet exist in the appropriate GS location
(currently we still check each buildspec -- a further opt would be to store the
most recent version processed or something similar in Firestore).
As part of this optimization, we also add the --force flag to allow the user
to force regeneration for existing buildspecs in GS.

This change includes refactors to per_project_buildspec_test.go to make
shared set up code more flexible.

See go/per-project-buildspecs and go/cros-partner-buildspec-sync for
context.

BUG=b:173383632
TEST=go test, manual runs

Change-Id: I970de10f38bf68c779e22e6fcb1615193fd84663
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3092717
Commit-Queue: Jack Neus <jackneus@google.com>
Reviewed-by: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: Sean McAllister <smcallis@google.com>
Cr-Commit-Position: refs/heads/main@{#43223}