Update dockcross images.
This includes an update to gcc-10 for manylinux-x64 images, and an
update to perl to get it working with the new compiler.

Because python puts the absolute path to the compiler into the
LDSHARED config variable, we override it in the environment to
point to the new toolchain. This will be necessary at least until
new cpython(3) packages are built. Other variables like CC, etc
are already overridden in the environment.

Change-Id: Ib4c156b963f9693be038172bc8a42facca0d128d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3106178
Reviewed-by: Owen Rodley <orodley@chromium.org>
Reviewed-by: Chenlin Fan <fancl@chromium.org>
Commit-Queue: Brian Ryner <bryner@google.com>
Cr-Commit-Position: refs/heads/main@{#43333}