Reorganize code to support getting exe from CAS.
In order to support using led to edit the recipe bundle, the
bootstrapper needs to be able to extract the CAS digest from the
led_recipe_bundle property and download the isolated from CAS rather
than getting the CIPD package.

This splits the exe bootstrapper so that it has an operation for getting
details about the executable that should be used and an operation for
setting up the exe and adds the bootstrapped exe as a property that will
be reported to the recipe so that it can expose it on the build. The
property bootstrapper is similarly split up into an operation to get the
build config, which will perform all of the work of reading the files
with the builder properties, but instead return a config object that has
an operation that takes the exe information and produces the properties
to use when executing the bootstrapped executable.

A follow-up CL will add the support for the CAS recipe bundle.

Bug: 1210566
Change-Id: I1c25ac00d02ef6013a17bbc903f167072d5a4ab6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3123463
Auto-Submit: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Garrett Beaty <gbeaty@google.com>
Reviewed-by: Yiwei Zhang <yiwzhang@google.com>
Cr-Commit-Position: refs/heads/main@{#43521}