Fix flaky tests
The tests were flaky because the side effect for GitilesRepository.GetSource() was set up through a list. But since the underlying caller code is multi-threaded, it isn't necessary that GetSource() be invoked in the order corresponding to return value defined in the side_effect. Changing side_effect to a function which returns based on the parameters recevied should solve it.

Bug: 1243962
Change-Id: I75d08956c6badfbb72d2832cbaf0e23a0bbbc0b1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3137333
Auto-Submit: Prakhar Asthana <pasthana@google.com>
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>
Commit-Queue: Zhaoyang Li <zhaoyangli@chromium.org>
Cr-Commit-Position: refs/heads/main@{#43526}