[result_adapter] Print payload's output.
This is so that this output is available as a step log.

Some developers have been surprised by the output of their tests
no longer being available at the output linked directly from the
step. This is available as a ResultDB artifact in the results
tab, but it is not immediately apparent.

With this change, the output of the tests remains in the output
of the step, while also being uploaded to ResultDB.

R=ddoman,chanli

Change-Id: Iac65a27f1a4035b5b07e0720f94916a9dd620178
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3133951
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Roberto Carrillo <robertocn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#43596}