[recovery]: Check if host is reachable via SSH
This commit is a part of the PARIS project. It provides an function to
check whether a host is reachable through SSH.

This function will be used in recovery actions such as determining
whether the path to USB drive if the servo-host is reachable.

BUG=b:195583889
TEST=build

Change-Id: I50331c8f9f3879b32fa517c7cb5c2c378e6f8e7a
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3145715
Commit-Queue: Otabek Kasimov <otabek@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#43678}