[recovery]: Verifier for Servo USB
This commit is a part of the PARIS project. It provides an
implementation of the servo USB detector.

This detector is the first step in the process of validating servo
USB. If the servo USB path is detected to be present, the next steps
such as verifying that USB drive is in good condition will be
handled. These subsequent steps will be included in separate CLs.

BUG=b:195583889
TEST=build, manual testing.

Change-Id: Ib4ca8c206c91f30adc6bc8db8f24dc8069ae3a0e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3150355
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#43690}