[recovery] Add Metrics API.
Metrics is the subset of the Karte API that is used by the recovery library.
This allows the recovery library to have more than one backing implementation
of the Karte API. The recovery library will have at least two implementations of
this API, a basic one that logs JSON at the debug level, and a more complex one
that calls Karte.

BUG=b/183724382
TEST=make test

Change-Id: I278fe47ab70047c6ad09e0dbbdebb94aaa463267
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3100488
Commit-Queue: Gregory NISBET <gregorynisbet@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#43704}