Weetbix: Add basic SPA UI
This converts the existing server rendered index page to a client
rendered page using lit-element.

This is loosely based on the MILO UI, but started from scratch, we will
add to it as required.

Note there are no tests in this CL as the test infrastructure requires a
lot of config/boilerplate.  This will be added in the next CL.

BUG=1243174

Change-Id: I75bbc7919b13ed93b52a7209967d08351f0aefb5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3149880
Commit-Queue: Matthew Warton <mwarton@google.com>
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Patrick Meiring <meiring@google.com>
Cr-Commit-Position: refs/heads/main@{#43714}