[karte] Remove validation comments from low-level conversion routines.
We have two conversion functions that convert between an enum and an int32.
Validating to make sure that the int32 is in range is not a good idea, karte
should keep going and record the event anyway even if the record is out of range.

BUG=b/183724382
TEST=make test

Change-Id: Ie3261cfb5ecf4a8fa6304e2c75f33f6f6098d3ac
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3150363
Reviewed-by: Xianuo Wang <xianuowang@chromium.org>
Commit-Queue: Gregory NISBET <gregorynisbet@google.com>
Cr-Commit-Position: refs/heads/main@{#43730}