CTP2: Add skeleton luciexe.
- Right now just reads input request and naively takes the first
allowed value of each DutCriterion to form dimensions for
scheduling on Swarming.

- Logdog Butler is needed to run locally. I had a discussion on
LUCI Users, and the luciexe libs may be changed to not require
this, so the binary can be run directly locally. For now, have
a helper script to run the Butler, and produce the binary-encoded
Build proto the luciexe expects as input.

BUG=b:188712224
TEST=CQ

Change-Id: I9b0eed6d12f462fe650a5becc22170d54fb197d1
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3155402
Commit-Queue: Andrew Lamb <andrewlamb@chromium.org>
Reviewed-by: Jaques Clapauch <jaquesc@google.com>
Reviewed-by: Sean McAllister <smcallis@google.com>
Cr-Commit-Position: refs/heads/main@{#43750}