[recovery]: Verifier for Servo USB
This commit is a part of the PARIS project. It provides an
implementation of an action to verify that USB drive is in working
condition.

This action depends on an earlier action that determines whether a USB
drive exists or not. It then applies a check on the detected USB
drive.

BUG=b:195583889
TEST=build, manual testing.

Change-Id: I8a41c841740b57933b954263d5fd87091f901235
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3154344
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Cr-Commit-Position: refs/heads/main@{#43754}