upload_debug_symbols: Unpack the tarball and save only the debug symbols
After the .tgz has been downloaded and uncompressed, we read through the
tarball and save the symbol files locally. After we write all the files
locally, we return a list pointing to each's path for config
generation.

Test coverage for this method involved creating a mock tarball and
having the method extract it's contents. We verified the results by
checking it against the expected results. Additional the builder was run
locally using a real debug symbol in GS and the results were hand
verified.

BUG=b:195981092,b:197010274
TEST=unittests; local execution

Change-Id: I6e660466bb136cf8b8c93d3fb1f837cdeb4e2fc2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3150307
Commit-Queue: Julio Hurtado <juahurta@google.com>
Reviewed-by: George Engelbrecht <engeg@google.com>
Reviewed-by: Jack Neus <jackneus@google.com>
Reviewed-by: Mike Nichols <mikenichols@chromium.org>
Cr-Commit-Position: refs/heads/main@{#43819}