Add task retry limit for feature coverage
A task can now be retried twice. A worker leases the task and executes in a try block. All exceptions are silently ignored. At the end of lease period(1 hour) the task would be returned back to the queue, and would be available for other workers to retry.

Bug: 1141021
Change-Id: Ie023417db4d36b4a5305f2c44100c0c10e64d290
Reviewed-on: https://chromium-review.googlesource.com/c/infra/infra/+/3180628
Auto-Submit: Prakhar Asthana <pasthana@google.com>
Reviewed-by: Zhaoyang Li <zhaoyangli@chromium.org>
Commit-Queue: Prakhar Asthana <pasthana@google.com>
Cr-Commit-Position: refs/heads/main@{#44116}